backend/requirements/requirements-base.txt
alembic==0.9.2
Babel==2.4.0
blinker==1.4
click==6.7
elasticsearch==2.4.1
elasticsearch-dsl==2.2.0
enum-compat==0.0.2
enum34==1.1.6
eventlet==0.20.1
Flask==1.0.2
Flask-Babel==0.11.2
Flask-Cors==3.0.3
Flask-DotEnv==0.1.1
Flask-Elasticsearch==0.2.5
Flask-JWT-Extended==1.1.0
Flask-Mail==0.9.1
Flask-Migrate==2.0.4
Flask-Principal==0.4.0
flask-restplus==0.10.1
Flask-SQLAlchemy==2.2
Flask-Uploads==0.2.1
greenlet==0.4.12
humanize==0.5.1
itsdangerous==0.24
Jinja2==2.9.6
langdetect==1.0.7
MarkupSafe==1.0
marshmallow==2.13.1
nose==1.3.7
pbr==3.1.1
psycopg2==2.7.1
py==1.4.32
PyJWT==1.4.2
python-dateutil==2.6.1
python-editor==1.0.3
python-termstyle==0.1.10
pytz==2017.2
simplekv==0.10.0
six==1.10.0
SQLAlchemy==1.2.7
SQLAlchemy-Utils==0.32.14
stevedore==1.26.0
urllib3==1.21.1
webargs==1.5.3
Werkzeug==0.12.2
Whoosh==2.7.4
